+++ models/iaf_psc_exp_multisynapse.cpp
@@ -122,7 +122,7 @@ iaf_psc_exp_multisynapse::Parameters_::get( Dictionary& d ) const
   d[ names::C_m ] = C_;
   d[ names::tau_m ] = Tau_;
   d[ names::t_ref ] = refractory_time_;
-  d[ names::n_synapses ] = n_receptors_();
+  d[ names::n_synapses ] = static_cast< long >( n_receptors_() );
   d[ names::has_connections ] = has_connections_;
   d[ names::tau_syn ] = tau_syn_;
 }
