+++ models/static_synapse.h
@@ -189,7 +189,7 @@ static_synapse< targetidentifierT >::get_status( Dictionary& d ) const
 
   ConnectionBase::get_status( d );
   d[ names::weight ] = weight_;
-  d[ names::size_of ] = sizeof( *this );
+  d[ names::size_of ] = static_cast< long >( sizeof( *this ) );
 }
 
 template < typename targetidentifierT >
